fix: address code review issues #5-#11
- Split jest-harness-setup.ts (525 lines) into harness/polyfills.ts,
  harness/jest-compat.ts, and harness/snapshots.ts
- Support Indian numbering (hi-IN, bn) in insertGroupingSeparator
  with 3+2+2 digit grouping pattern
- Add notify-failure job to CI workflow that creates GitHub issues
  on nightly harness test failures
- Add build-time console.warn when babel plugin strips jest.mock()
  without factory (auto-mock not supported in harness)
- Simplify containsPunycode to single return expression
- Rename smoke test describe to "Harness Infrastructure Validation"
  to clarify it validates the harness pipeline, not Hermes features
- Add fast path in countLeadingZeroDecimals to skip toFixed() for
  values >= 1
